backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.3.0
groq==0.9.0
pandas==2.2.2
numpy>=1.26.4
scipy>=1.14.1
scikit-learn>=1.5.0
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
sqlmodel==0.0.19
reportlab==4.2.0
Pillow>=10.3.0
trafilatura>=1.9.0
beautifulsoup4==4.12.3
lxml>=5.2.2
chardet==5.2.0
openpyxl==3.1.3
loguru==0.7.2
slowapi==0.1.9
PyMuPDF>=1.24.4
python-jose[cryptography]==3.3.0
